Started new lobby view, improved mobile view, fixed some issues
- Fixed alerts hiding behind card
- Made mobile view look better (couldn't fix background yet)

Still need to remake LobbySettings
