+++ src/components/ui/ActivePlayersList.js
@@ -14,9 +14,9 @@ const PlayerAvatar = ({ name, color, isAdmin }) => {
     display: 'inline-flex',
     alignItems: 'center',
     justifyContent: 'center',
-    width: '80px',
-    height: '80px',
-    fontSize: '28px',
+    width: '60px',
+    height: '60px',
+    fontSize: '20px',
     fontWeight: '600',
     color: 'white',
     border: `2px solid ${backgroundColor}`,
