+++ src/styles/views/Home.scss
@@ -13,6 +13,7 @@
     border-radius: 10px;
     padding: 2rem;
     box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+    margin-top:1em;
   }
   &.title {
     &:hover {
@@ -82,7 +83,8 @@
   }
   &.alert{
     position: absolute;
-    top: 20px;
+    top: 40px;
     right: 0px;
+    z-index: 1;
   }
 }
