+++ src/styles/views/Lobby.scss
@@ -7,6 +7,7 @@
     justify-content: center;
     align-items: center;
     min-height: 100vh;
+    margin-top:1em;
   }
   &.card {
     background-color: rgba(255, 255, 255, 0.4);
@@ -14,11 +15,14 @@
     padding: 2rem;
     box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
     color: black;
+    margin-bottom:1em;
+    &.code {
+      margin-top:13px;
+    }
   }
-
   &.alert{
     position: absolute;
-    top: 40px;
+    top: 0px;
     right: 0px;
   }
   &.title {
@@ -91,19 +95,20 @@
       background: #da0000;
       border: solid #af0000;
     }
-    &.leave-btn{
-      &:active {
-        color: white !important;
-        border: solid #da0000;
-        background: #da0000;
-      }
-      position: absolute;
-      top: 20px;
-      right: 20px;
-      background: #da0000;
+  }
+  &.leave-btn{
+    &:hover {
+      background-color: #af0000;
       border: solid #af0000;
-      width:auto;
     }
-
+    &:active {
+      color: white !important;
+      border: solid #da0000;
+      background: #da0000;
+    }
+    background: #da0000;
+    border: solid #af0000;
+    width: 90%;
+    margin:auto;
   }
 }
